+++ LTEV2X/IMTA.h
@@ -16,90 +16,90 @@ public:
 	static const unsigned char m_scbySubPathNum = 20;
 	static const unsigned char m_scbyMidPathNum = 3;
 	//相关系数矩阵
-	static const float m_sacfConstantInHLoS[25];
-	static const float m_sacfConstantInHNLoS[25];
-	static const float m_sacfConstantUMiLoS[25];
-	static const float m_sacfConstantUMiNLoS[25];
-	static const float m_sacfConstantUMiO2I[25];
-	static const float m_sacfConstantSMaLoS[25];
-	static const float m_sacConstantSMaNLoS[25];
-	static const float m_sacfConstantUMaLoS[25];
-	static const float m_sacfConstantUMaNLoS[25];
-	static const float m_sacfConstantRMaLoS[25];
-	static const float m_sacfConstantRMaNLoS[25];
+	static const double m_sacfConstantInHLoS[25];
+	static const double m_sacfConstantInHNLoS[25];
+	static const double m_sacfConstantUMiLoS[25];
+	static const double m_sacfConstantUMiNLoS[25];
+	static const double m_sacfConstantUMiO2I[25];
+	static const double m_sacfConstantSMaLoS[25];
+	static const double m_sacConstantSMaNLoS[25];
+	static const double m_sacfConstantUMaLoS[25];
+	static const double m_sacfConstantUMaNLoS[25];
+	static const double m_sacfConstantRMaLoS[25];
+	static const double m_sacfConstantRMaNLoS[25];
 	//角度偏移数组
-	static const float m_sacfAngleOffset[m_scbySubPathNum];
-	static const float m_sacfMidPathDelayOffset[m_scbyMidPathNum];
+	static const double m_sacfAngleOffset[m_scbySubPathNum];
+	static const double m_sacfMidPathDelayOffset[m_scbyMidPathNum];
 	static const unsigned char m_sacbyMidPathIndex[m_scbySubPathNum];
 
 	//信道所需基本常量
-	float m_fAntGain;
-	float m_fMaxAttenu; // dBm
+	double m_fAntGain;
+	double m_fMaxAttenu; // dBm
 	unsigned char m_byTxAntNum;
 	unsigned char m_byRxAntNum;
-	float * m_pfTxSlantAngle; // degree
-	float * m_pfRxSlantAngle; // degree
-	float * m_pfTxAntSpacing;
-	float * m_pfRxAntSpacing;
-	float m_fTxAngle;
-	float m_fRxAngle;
+	double * m_pfTxSlantAngle; // degree
+	double * m_pfRxSlantAngle; // degree
+	double * m_pfTxAntSpacing;
+	double * m_pfRxAntSpacing;
+	double m_fTxAngle;
+	double m_fRxAngle;
 
 	//信道所需配置常量
 	unsigned short m_byPathNum;
-	float m_fVelocity;
-	float m_fvAngle;
-	float m_fC;
-	float m_fPathShadowSTD;
-	float m_fAoDRatio;
-	float m_fAoARatio;
-	float m_fXPR;
-	float m_fDS;
-	float m_fDSRatio;
-	float m_fAoD;
-	float m_fAoA;
-	float m_fK;
-	float m_fKDB;
-	float m_fD;
+	double m_fVelocity;
+	double m_fvAngle;
+	double m_fC;
+	double m_fPathShadowSTD;
+	double m_fAoDRatio;
+	double m_fAoARatio;
+	double m_fXPR;
+	double m_fDS;
+	double m_fDSRatio;
+	double m_fAoD;
+	double m_fAoA;
+	double m_fK;
+	double m_fKDB;
+	double m_fD;
 
 
 	bool m_bLoS;
 	bool m_bBuilt;
 	bool m_bEnable;
-	float m_fPLSF;
+	double m_fPLSF;
 
 	unsigned char m_byPathFirst;
 	unsigned char m_byPathSecond;
 	//信道所需存储常量
-	float *m_pfGain;
-	float *m_pfSinAoD;
-	float *m_pfCosAoD;
-	float *m_pfSinAoA;
-	float *m_pfCosAoA;
-	float *m_pfPhase;
+	double *m_pfGain;
+	double *m_pfSinAoD;
+	double *m_pfCosAoD;
+	double *m_pfSinAoA;
+	double *m_pfCosAoA;
+	double *m_pfPhase;
 
-	float m_fGainLoS;
-	float m_fSinAoDLoS;
-	float m_fSinAoALoS;
-	float m_fCosAoALoS;
-	float *m_pfPhaseLoS;
+	double m_fGainLoS;
+	double m_fSinAoDLoS;
+	double m_fSinAoALoS;
+	double m_fCosAoALoS;
+	double *m_pfPhaseLoS;
 
 	//FFT所用变量
 	unsigned short m_wFFTNum;
 	unsigned char m_byFFTOrder;
-	float m_fFFTTime;
+	double m_fFFTTime;
 	unsigned short m_wHNum;
 	unsigned short *m_pwFFTIndex;
 public:
 	cIMTA(void);
 	~cIMTA(void);
-	bool Build(float* t_Pl, float t_fFrequency/*Hz*/, sLocation &t_eLocation, sAntenna &t_eAntenna,  float t_fVelocity/*km/h*/, float t_fVAngle/*degree*/);
+	bool Build(double* t_Pl, double t_fFrequency/*Hz*/, sLocation &t_eLocation, sAntenna &t_eAntenna,  double t_fVelocity/*km/h*/, double t_fVAngle/*degree*/);
 	bool Enable(bool *t_pbEnable);
-	void Calculate(float* t_HAfterFFT, float t_fT/*s*/, float *t_pfTemp, float *t_pfSin, float *t_pfCos, float *t_pfH, float *t_pfHFFT);
-	float GetPLSF(void)
+	void Calculate(double* t_HAfterFFT, double t_fT/*s*/, double *t_pfTemp, double *t_pfSin, double *t_pfCos, double *t_pfH, double *t_pfHFFT);
+	double GetPLSF(void)
 	{
 		return m_fPLSF;
 	}
-    float GetAntGain(void)
+    double GetAntGain(void)
 	{
 		return m_fAntGain;
 	}
