+++ LTEV2X/RSU.h
@@ -18,11 +18,11 @@ public:
 	std::list<int> m_VeUEIdList;//当前RSU范围内的VeUEId编号容器,RRM_DRA模块需要
 	int m_DRAClusterNum;//一个RSU覆盖范围内的簇的个数,RRM_DRA模块需要
 	std::vector<std::list<int>> m_DRAClusterVeUEIdList;//存放每个簇的VeUE的Id的容器,下标代表簇的编号
-	float m_fAbsX;
-	float m_fAbsY;
+	double m_fAbsX;
+	double m_fAbsY;
 
 	cIMTA *imta;
-	float m_fantennaAngle;
+	double m_fantennaAngle;
 
 	/*------------------成员函数------------------*/
 	void initializeUrban(sRSUConfigure &t_RSUConfigure);
